Add exception handler and response utility
