+++ src/main/java/net/smartcosmos/cluster/userdetails/service/AuthenticateUserServiceDefault.java
@@ -8,10 +8,11 @@ import org.springframework.security.core.AuthenticationException;
 import org.springframework.stereotype.Service;
 
 import net.smartcosmos.cluster.userdetails.domain.AuthenticateUserRequest;
-import net.smartcosmos.cluster.userdetails.domain.MessageResponse;
 import net.smartcosmos.cluster.userdetails.domain.UserDetails;
 
-import static net.smartcosmos.cluster.userdetails.domain.MessageResponse.CODE_ERROR;
+import static net.smartcosmos.cluster.userdetails.util.ResponseEntityFactory.invalidDataReturned;
+import static net.smartcosmos.cluster.userdetails.util.ResponseEntityFactory.invalidUsernameOrPassword;
+import static net.smartcosmos.cluster.userdetails.util.ResponseEntityFactory.success;
 
 @Slf4j
 @Service
@@ -35,16 +36,13 @@ public class AuthenticateUserServiceDefault implements AuthenticateUserService {
 
             if (userDetailsService.isValid(userDetails)) {
                 log.info("Validation of authentication response for user {} : valid", request.getName());
-                return ResponseEntity.ok(userDetails);
+                return success(userDetails);
             }
             log.info("Validation of authentication response for user {} : invalid", request.getName());
-            return ResponseEntity.badRequest()
-                .body(new MessageResponse(CODE_ERROR, "Invalid data returned"));
+            return invalidDataReturned();
         } catch (AuthenticationException e) {
             log.info("Authenticating user {} failed. Request was {}. Exception: {}", request.getName(), request, e.toString());
-            MessageResponse messageResponse = new MessageResponse(CODE_ERROR, e.getMessage());
-            return ResponseEntity.badRequest()
-                .body(messageResponse);
+            return invalidUsernameOrPassword();
         }
     }
 }
